Dodane klase koje nedostaju i popravljeni neki bagovi, testirana aplikacija, valjda radi

+++ src/ba/unsa/etf/rpr/Artikl.java
@@ -4,7 +4,7 @@ public class Artikl {
     String kod = new String("");
     String naziv = new String("");
 
-    Artikl(String naziv, double cijena, String kod) {
+    Artikl(String naziv, int cijena, String kod) {
         this.naziv = naziv;
         this.cijena = cijena;
         this.kod=kod;
@@ -14,7 +14,7 @@ public class Artikl {
         return kod;
     }
 
-    public double getCijena() {
+    public int getCijena() {
         return cijena;
     }
